Factorize some view code
