+++ include/view.h
@@ -208,6 +208,8 @@ int rofi_view_error_dialog ( const char *msg, int markup  );
  */
 void rofi_view_queue_redraw ( void );
 
+void rofi_view_calculate_window_position ( RofiViewState *state );
+
 /**
  * Cleanup internal data of the view.
  */
@@ -315,26 +317,22 @@ void rofi_view_get_size ( RofiViewState * state, gint *width, gint *height );
 
 typedef struct _view_proxy {
     RofiViewState* (*create) ( Mode *sw, const char *input, MenuFlags menu_flags, void ( *finalize )( RofiViewState * ) );
-    void (*finalize) ( RofiViewState *state );
-    MenuReturn (*get_return_value) ( const RofiViewState *state );
-    unsigned int (*get_next_position) ( const RofiViewState *state );
     void (*handle_text) ( RofiViewState *state, char *text );
-    void (*handle_mouse_motion) ( RofiViewState *state, gint x, gint y );
     void (*maybe_update) ( RofiViewState *state );
     void (*temp_configure_notify) ( RofiViewState *state, xcb_configure_notify_event_t *xce );
     void (*temp_click_to_exit) ( RofiViewState *state, xcb_window_t target );
     void (*frame_callback) ( void );
-    unsigned int (*get_completed) ( const RofiViewState *state );
     const char * (*get_user_input) ( const RofiViewState *state );
     void (*set_selected_line) ( RofiViewState *state, unsigned int selected_line );
-    unsigned int (*get_selected_line) ( const RofiViewState *state );
-    void (*restart) ( RofiViewState *state );
     gboolean (*trigger_action) ( RofiViewState *state, BindingsScope scope, guint action );
     void (*free) ( RofiViewState *state );
     RofiViewState * (*get_active) ( void );
     void (*set_active) ( RofiViewState *state );
     int (*error_dialog) ( const char *msg, int markup  );
     void (*queue_redraw) ( void );
+    void (*calculate_window_position) ( RofiViewState *state );
+    int (*calculate_window_height) ( RofiViewState *state );
+    void (*window_update_size) ( RofiViewState *state );
 
     void (*cleanup) ( void );
     Mode * (*get_mode) ( RofiViewState *state );
@@ -349,8 +347,6 @@ typedef struct _view_proxy {
     void (*workers_finalize) ( void );
     void (*get_current_monitor) ( int *width, int *height );
     void (*capture_screenshot) ( void );
-
-    void (*ellipsize_start) ( RofiViewState *state );
 
     void (*set_size) ( RofiViewState * state, gint width, gint height );
     void (*get_size) ( RofiViewState * state, gint *width, gint *height );
